重新排版代码 去掉无用的代码

+++ src/bullet_settings/rotate_flower.hpp
@@ -6,8 +6,8 @@ namespace game::bullet_settings{
 class RotateFlower : public game::BulletManager{
   GDCLASS(RotateFlower, game::BulletManager)
 private:
-    int frame_counter = 0; // 帧率计时
-    float current_base_degree = 0.0f; // 记录当前转到了多少度
+  int frame_counter = 0; // 帧率计时
+  double current_base_degree = 0.0f; // 记录当前转到了多少度
 public:
   void shoot() override;
   static void _bind_methods(){};
